+++ src/index.js
@@ -43,15 +43,24 @@ const run = async () => {
             process.exit(0);
         }
 
+        const [branchName, project] = branch.split('/');
+        const isReleaseBranch = releaseBranches.find((releaseBranch) => branchName.startsWith(releaseBranch));
+
         const isExcluded = wcmatch(exclude ? exclude.split(',')
             .map((name) => name.trim()) : []);
 
-        const modifiedFiles = (await getModifiedFiles({
+        let modifiedFiles = (await getModifiedFiles({
             octokit,
             repo,
             owner,
             pullNumber,
         })).filter((file) => !isExcluded(file));
+
+        // Limits modified files to the ones in the project folder of the release
+        if (isReleaseBranch && project) {
+            const isInAppFolder = wcmatch([`projects/${project}/**`, `packages/${project}/**`]);
+            modifiedFiles = modifiedFiles.filter((file) => isInAppFolder(file));
+        }
 
         const folders = await getFolders(sources);
 
@@ -71,14 +80,10 @@ const run = async () => {
             validateChangelog(changelogContent);
         }));
 
-        // Checks if the branch is release or once of release_branches input.
-        if (releaseBranches.find((releaseBranch) => branch.startsWith(releaseBranch))) {
+        if (isReleaseBranch) {
             const changelogs = modifiedFiles.filter((file) => file.endsWith('CHANGELOG.md'));
 
             if (changelogs.length) {
-                /** If branch name contains project ex: release/account */
-                const project = branch.includes('/') && branch.split('/').slice(-1)[0];
-
                 if (project) {
                     const projectChangelog = changelogs.find((file) => file.includes(`${project}/CHANGELOG.md`));